builder: split nix instantiation and build
The build is split into a two-step process:
- nix-instantiate, which is called with -vv and parses its stderr (and
  produces drv files)
- nix-build, which builds the drv file

The idea behind this is that we can’t easily write tests which use
`nix-build` (nix-in-nix problem), so we are only gonna use the first
step for tests (which is enough to check the stderr parsing stuff).

Another problem this solves is that the stderr of nix-build is now
separated from the stderr of nix-instantiate, and nix-build is not
called with -vv, meaning we can show it to the user while it’s
generated, making it possible to fix the silence problem lorri has.

Closes https://github.com/target/lorri/issues/126