nix/ci/sandbox: move to a unshare/bindmount/chroot sandbox script
buildSandbox was wrapping the executable with a C program, and by
mounting the full nix store read/write we defeated the whole idea
behind it in the first place, which was to only mount a subset of the
store.

So instead there is a simple strategy to achieve a simple sandbox:

1. unshare the file namespace
2. bind-mount /nix /sys /proc etc. into a sandbox directory
3. chroot into the sandbox directory

different sandboxes won’t see each other, the sandboxed command can
work like it’s root, and nix handle all paths for us.