release.nix/default.nix: always use the relative src
We shouldn’t pass src explicitely (because it can lead to a mismatch
between `default.nix` and source code), and indeed we never need to,
not even for `release.nix` and `lorri self-upgrade`.

We have to add an `isFunction` check, otherwise `self-upgrade` fails
for branches where the `src` attribute is still passed manually to
`release.nix`.